Enabled some consistency checks in the gc when compiled without rtldebug (only activated on debug level 1 or higher). Always define _verify_internals to be able to use this.
Made it possible to turn on trace messages for the gc only with
trace(1,"gc"). This is the embryo of a facility based trace system. Still to
do: Raise all the global trace levels to make room for gc only trace at
level 1, and fix a framework for trace facilities.