monger will now attempt to delete the previous version of a module, if it exists, is tagged with __version (indicating it's monger ready), and if it contains an array __components that contains a list of files and directories belonging to this module (relative to the module root).
it's sort of a "running with scissors" approach because:

a) the module developer must provide a list of constituent components
b) the module developer must be trusted to provide a) that is accurate.

we mitigate the danger of b) by not allowing files outside of the
module directory to be deleted and by not recursively deleting files. that
means that if a file is left over, we don't delete it and its containing
directory structure.

i'd be willing to hear alternate ideas concerning the need for cleanly deleting
module components.