Reverted automatic switching of character_set_connection since it isn't side effect free.
E.g. the expression 'ü' < 'v' would evaluate differently depending on
whether the query is sent as latin1 or utf-8.

Another example:

  insert into foo values ('<text data>', '<binary data>');

Here <binary data> would be clobbered if <text data> happen to contain
wide chars.

So it's simply not possible to get transparent unicode support in
MySQL. Deal with it. :\

Also tuned the MYSQL_CHARSET_DEBUG a little.