Courtesy of Betrand LUPART:
Here's a patch for having a more clear error message and abort if the
specs file isn't present at the expected location when running "pike -x
monger --install" for example.

The current error is actually a user-unfriendly backtrace.

Maybe the check for this file could be moved elsewhere, since before
checking for the specs file, monger select the good version of the
software to install and downloads it.