Changed the environment implementation to use __builtin._getenv and __builtin._putenv so that the real environment is queried and changed. There is still a local mapping for caching purposes, but it has a different format.
The _main function neither receives nor passes along an environment
mapping to the script main() or to the --execute code.

NOTE: Both of the above have compat implications, so compat fallbacks
have been added.

Also added a kludge to the master_efuns stuff to use the versions from
the right compat version. That's not entirely right, but it's a bit
better at least.