Disabled renegotiation, to address the ssl/tls renegotiate MITM attack.
See http://web.nvd.nist.gov/view/vuln/detail?vulnId=CVE-2009-3555.

Secure renegotiation according to the tls extension remains to be
implemented (the spec is still a draft at the time of this writing). Note
that renegotiation often didn't work anyway due to a bug, so this change
might not be much of a compatibility issue anyway.